migrate delete account types api
