+++ client/src/hooks/mutations/accountTypes/useDeleteAccountTypeMutation.ts
@@ -1,0 +1,34 @@
+import { notifications } from "@mantine/notifications";
+import { useMutation, useQueryClient } from "@tanstack/react-query";
+import { AxiosError } from "axios";
+import {
+  accountsQueryKey,
+  accountTypesQueryKey,
+  institutionsQueryKey,
+  translateAxiosError,
+} from "~/helpers/requests";
+import { useAuth } from "~/providers/AuthProvider/AuthProvider";
+
+export const useDeleteAccountTypeMutation = () => {
+  const queryClient = useQueryClient();
+  const { request } = useAuth();
+
+  return useMutation({
+    mutationFn: async (guid: string) =>
+      await request({
+        url: "/api/accountType",
+        method: "DELETE",
+        params: { guid },
+      }),
+    onSuccess: async () => {
+      await queryClient.invalidateQueries({ queryKey: [accountTypesQueryKey] });
+      await queryClient.invalidateQueries({ queryKey: [institutionsQueryKey] });
+      await queryClient.invalidateQueries({ queryKey: [accountsQueryKey] });
+    },
+    onError: (error: AxiosError) =>
+      notifications.show({
+        color: "var(--button-color-destructive)",
+        message: translateAxiosError(error),
+      }),
+  });
+};
